Modify HConverter to provide 'ready' functions
These can be called to signal that a result of a conversion is available.
It is required by the architechture that a value given as parameter to
a 'ready' callback do not change untill the next Write() - after which the
value may become undefined - thereofre, it is important that the 'ready'
callback executes fast and, if needed, copies the provided data to its own
persistent buffers.