Add first (proof of concept) example - a simple saq (Grimeton) receiver
This commit included a 1.3MB wav file with a recording of the antenna signal
during saq's transmission - normally no such files should be committed, however
this is needed for the example, and reception (and decoding) of this signal is
the prime driver behind the creation of the Hardt framework.
So..  "ok this time, but dont do it again! ;)"